# Added BST for efficient deletion