--- a/cache_test.go
+++ b/cache_test.go
@@ -1591,3 +1591,19 @@ func BenchmarkRWMutexMapSetDeleteSingleLock(b *testing.B) {
 		mu.Unlock()
 	}
 }
+
+func BenchmarkDeleteExpired(b *testing.B) {
+	b.StopTimer()
+	tc := New(1, 0)
+	tc.mu.Lock()
+	for i := 0; i < b.N; i++ {
+		tc.set(strconv.Itoa(i), "bar", DefaultExpiration)
+	}
+	tc.mu.Unlock()
+	time.Sleep(100)
+	if _, found := tc.Get("0"); found {
+		b.Fatal("0 found")
+	}
+	b.StartTimer()
+	tc.DeleteExpired()
+}