@pablodz 2024-01-03
Found this fork: https://github.com/arp242/zcache